Changed alot, like A LOT, I made cornucopia-cards.json which contains all the OWASP cornucopia cards, I dont own the license to these ofcourse, and this is ofc licensed in accordance the guidelines presented in the original license, i'll put it in the readme when i get to it.
I created a basic flow for card dealing to a room, only one room exists now, I would like this to be dynamic, so there potentially could be more rooms.
When a room is joined, there is basic functionality for dealing cards, when cards are dealt, the room is closed for joining, handling the unsuccessful joins is not done yet.

-There is a lot of redundancy in the code, and it needs cleaning atm, like, redux is implemented, but currently it isn't used for anything, but I might have a use for it later, so I wont remove it. yet.
-scoreboards is not done .
-Dynamic room generation is not implemented